+++ migration/src/main/java/uk/ac/bbsrc/tgac/miso/migration/destination/MisoServiceManager.java
@@ -55,7 +55,6 @@ import uk.ac.bbsrc.tgac.miso.persistence.impl.HibernateTissueMaterialDao;
 import uk.ac.bbsrc.tgac.miso.persistence.impl.HibernateTissueOriginDao;
 import uk.ac.bbsrc.tgac.miso.persistence.impl.HibernateTissueTypeDao;
 import uk.ac.bbsrc.tgac.miso.service.BoxService;
-import uk.ac.bbsrc.tgac.miso.service.DefaultLibraryDesignCodeService;
 import uk.ac.bbsrc.tgac.miso.service.QualityControlService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultBoxService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultChangeLogService;
@@ -64,6 +63,7 @@ import uk.ac.bbsrc.tgac.miso.service.impl.DefaultExperimentService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultIndexService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultKitService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultLabService;
+import uk.ac.bbsrc.tgac.miso.service.impl.DefaultLibraryDesignCodeService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultLibraryDesignService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultLibraryDilutionService;
 import uk.ac.bbsrc.tgac.miso.service.impl.DefaultLibraryService;
