+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultPartitionQCService.java
@@ -13,6 +13,8 @@ import uk.ac.bbsrc.tgac.miso.core.data.PartitionQCType;
 import uk.ac.bbsrc.tgac.miso.core.data.Run;
 import uk.ac.bbsrc.tgac.miso.core.store.PartitionQcStore;
 import uk.ac.bbsrc.tgac.miso.service.ContainerService;
+import uk.ac.bbsrc.tgac.miso.service.PartitionQCService;
+import uk.ac.bbsrc.tgac.miso.service.RunService;
 import uk.ac.bbsrc.tgac.miso.service.security.AuthorizationManager;
 
 @Service
